requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.2.3
PyPDF2==3.0.1
docx2txt==0.8
sentence-transformers==2.2.2
torch==2.0.1
requests==2.31.0
beautifulsoup4==4.12.2
